Change from Object.create() to Object.assign()
Object.create() creates a new object, assigning the properties of argument object to prototype. This prototype is lost if the object is passed in the response of express or any HTTP server response. Object.assign() will retain all properties in the new object which can be easily serialized via JSON and sent as a response.